requirements.txt
arrow
prometheus_client
python3-requests